Change zero_adapter to ze_graph_ext_wrapper
