+++ src/plugins/intel_npu/src/backend/src/zero_backend.cpp
@@ -14,31 +14,31 @@ namespace intel_npu {
 ZeroEngineBackend::ZeroEngineBackend(const Config& config) : _logger("ZeroEngineBackend", Logger::global().level()) {
     _logger.debug("ZeroEngineBackend - initialize started");
 
-    _instance = std::make_shared<ZeroInitStructsHolder>();
+    _initStruct = std::make_shared<ZeroInitStructsHolder>();
 
-    auto device = std::make_shared<ZeroDevice>(_instance);
+    auto device = std::make_shared<ZeroDevice>(_initStruct);
     _devices.emplace(std::make_pair(device->getName(), device));
     _logger.debug("ZeroEngineBackend - initialize completed");
 }
 
 uint32_t ZeroEngineBackend::getDriverVersion() const {
-    return _instance->getDriverVersion();
+    return _initStruct->getDriverVersion();
 }
 
 uint32_t ZeroEngineBackend::getGraphExtVersion() const {
-    return _instance->getGraphDdiTable().version();
+    return _initStruct->getGraphDdiTable().version();
 }
 
 bool ZeroEngineBackend::isBatchingSupported() const {
-    return _instance->isExtensionSupported("ZE_extension_graph_1_6", ZE_MAKE_VERSION(1, 6));
+    return _initStruct->isExtensionSupported("ZE_extension_graph_1_6", ZE_MAKE_VERSION(1, 6));
 }
 
 bool ZeroEngineBackend::isCommandQueueExtSupported() const {
-    return _instance->isExtensionSupported(std::string(ZE_COMMAND_QUEUE_NPU_EXT_NAME), ZE_MAKE_VERSION(1, 0));
+    return _initStruct->isExtensionSupported(std::string(ZE_COMMAND_QUEUE_NPU_EXT_NAME), ZE_MAKE_VERSION(1, 0));
 }
 
 bool ZeroEngineBackend::isLUIDExtSupported() const {
-    return _instance->isExtensionSupported(std::string(ZE_DEVICE_LUID_EXT_NAME), ZE_MAKE_VERSION(1, 0));
+    return _initStruct->isExtensionSupported(std::string(ZE_DEVICE_LUID_EXT_NAME), ZE_MAKE_VERSION(1, 0));
 }
 
 ZeroEngineBackend::~ZeroEngineBackend() = default;
@@ -69,23 +69,11 @@ const std::vector<std::string> ZeroEngineBackend::getDeviceNames() const {
 }
 
 void* ZeroEngineBackend::getContext() const {
-    return _instance->getContext();
+    return _initStruct->getContext();
 }
 
-void* ZeroEngineBackend::getDriverHandle() const {
-    return _instance->getDriver();
-}
-
-void* ZeroEngineBackend::getDeviceHandle() const {
-    return _instance->getDevice();
-}
-
-ze_graph_dditable_ext_curr_t& ZeroEngineBackend::getGraphDdiTable() const {
-    return _instance->getGraphDdiTable();
-}
-
-ze_command_queue_npu_dditable_ext_curr_t& ZeroEngineBackend::getCommandQueueDdiTable() const {
-    return _instance->getCommandQueueDdiTable();
+const std::shared_ptr<ZeroInitStructsHolder>& ZeroEngineBackend::getInitStruct() const {
+    return _initStruct;
 }
 
 void ZeroEngineBackend::updateInfo(const Config& config) {
